python-django/requirements.txt
Django==5.2
django-cors-headers==4.7.0
django-extensions==4.1
psycopg==3.2.6
python-dateutil==2.9.0.post0
pytz==2025.2
